Move point kernel implementation to medium.
