+++ core/specfem/point/kernels.hpp
@@ -1,329 +1,13 @@
 #pragma once
-#include "impl/point_container.hpp"
-#include <Kokkos_Core.hpp>
 
-namespace specfem {
-namespace point {
+#include "medium/dim2/acoustic/isotropic/kernels.hpp"
+#include "medium/dim2/elastic/anisotropic/kernels.hpp"
+#include "medium/dim2/elastic/isotropic/kernels.hpp"
+#include "medium/dim2/elastic/isotropic_cosserat/kernels.hpp"
+#include "medium/dim2/electromagnetic/isotropic/kernels.hpp"
+#include "medium/dim2/poroelastic/isotropic/kernels.hpp"
 
-namespace impl {
-namespace kernels {
-
-/**
- * @defgroup specfem_point_kernels_elastic_isotropic 2D Elastic Isotropic
- * Kernels
- * @{
- */
-/**
- * @ingroup specfem_point_kernels_elastic_isotropic
- * @brief Data container to hold misfit kernels of 2D elastic isotropic media at
- * a quadrature point
- *
- * @tparam UseSIMD Boolean indicating whether to use SIMD intrinsics
- *
- * <b> Accessor functions generated by the </b> \c POINT_CONTAINER \c MACRO
- *
- * - Get @f$ K_{\rho} @f$
- *   @code KOKKOS_INLINE_FUNCTION const value_type rho() const @endcode
- * - Get @f$ K_{\mu} @f$
- *   @code KOKKOS_INLINE_FUNCTION const value_type mu() const @endcode
- * - Get @f$ K_{\kappa} @f$
- *   @code KOKKOS_INLINE_FUNCTION const value_type kappa() const @endcode
- * - Get @f$ K_{\rho^\prime} @f$
- *   @code KOKKOS_INLINE_FUNCTION const value_type rhop() const @endcode
- * - Get @f$ K_{\alpha} @f$
- *   @code KOKKOS_INLINE_FUNCTION const value_type alpha() const @endcode
- * - Get @f$ K_{\beta} @f$
- *   @code KOKKOS_INLINE_FUNCTION const value_type beta() const @endcode
- *
- */
-template <specfem::dimension::type DimensionTag,
-          specfem::element::medium_tag MediumTag, bool UseSIMD>
-struct data_container<
-    DimensionTag, MediumTag, specfem::element::property_tag::isotropic, UseSIMD,
-    std::enable_if_t<specfem::element::is_elastic<MediumTag>::value> >
-    : public KernelsAccessor<DimensionTag, MediumTag,
-                             specfem::element::property_tag::isotropic,
-                             UseSIMD> {
-
-  using base_type =
-      KernelsAccessor<DimensionTag, MediumTag,
-                      specfem::element::property_tag::isotropic, UseSIMD>;
-
-  using value_type = typename base_type::value_type;
-  using simd = typename base_type::simd;
-
-  POINT_CONTAINER(rho, mu, kappa, rhop, alpha, beta)
-};
-/** @} */ // end of group
-
-/**
- * @defgroup specfem_point_kernels_dim2_elastic_anisotropic 2D Elastic
- * Anisotropic Kernels
- * @{
- */
-
-/**
- * @ingroup specfem_point_kernels_dim2_elastic_anisotropic
- * @brief Data container to hold misfit kernels of 2D elastic anisotropic media
- * at a quadrature point
- *
- * @tparam UseSIMD Boolean indicating whether to use SIMD intrinsics
- *
- * <b> Accessor functions generated by the </b> \c POINT_CONTAINER \c MACRO
- *
- * - Get @f$ K_{\rho} @f$
- *   @code KOKKOS_INLINE_FUNCTION const value_type rho() const @endcode
- * - Get @f$ K_{c_{11}} @f$
- *   @code KOKKOS_INLINE_FUNCTION const value_type c11() const @endcode
- * - Get @f$ K_{c_{13}} @f$
- *   @code KOKKOS_INLINE_FUNCTION const value_type c13() const @endcode
- * - Get @f$ K_{c_{15}} @f$
- *   @code KOKKOS_INLINE_FUNCTION const value_type c15() const @endcode
- * - Get @f$ K_{c_{33}} @f$
- *   @code KOKKOS_INLINE_FUNCTION const value_type c33() const @endcode
- * - Get @f$ K_{c_{35}} @f$
- *   @code KOKKOS_INLINE_FUNCTION const value_type c35() const @endcode
- * - Get @f$ K_{c_{55}} @f$
- *   @code KOKKOS_INLINE_FUNCTION const value_type c55() const @endcode
- */
-template <specfem::element::medium_tag MediumTag, bool UseSIMD>
-struct data_container<
-    specfem::dimension::type::dim2, MediumTag,
-    specfem::element::property_tag::anisotropic, UseSIMD,
-    std::enable_if_t<specfem::element::is_elastic<MediumTag>::value> >
-    : public KernelsAccessor<specfem::dimension::type::dim2, MediumTag,
-                             specfem::element::property_tag::anisotropic,
-                             UseSIMD> {
-  using base_type =
-      KernelsAccessor<specfem::dimension::type::dim2, MediumTag,
-                      specfem::element::property_tag::anisotropic, UseSIMD>;
-
-  using value_type = typename base_type::value_type;
-  using simd = typename base_type::simd;
-
-  POINT_CONTAINER(rho, c11, c13, c15, c33, c35, c55)
-};
-
-/** @} */ // end of group
-
-/**
- * @defgroup specfem_point_kernels_dim2_elastic_isotropic_cosserat 2D Cosserat
- * Kernels
- * @{
- */
-
-/**
- * @brief Placeholder container for 2D elastic isotropic cosserat media
- *
- * @tparam MediumTag
- * @tparam UseSIMD
- */
-template <specfem::element::medium_tag MediumTag, bool UseSIMD>
-struct data_container<
-    specfem::dimension::type::dim2, MediumTag,
-    specfem::element::property_tag::isotropic_cosserat, UseSIMD,
-    std::enable_if_t<specfem::element::is_elastic<MediumTag>::value> >
-    : public KernelsAccessor<specfem::dimension::type::dim2, MediumTag,
-                             specfem::element::property_tag::isotropic_cosserat,
-                             UseSIMD> {
-
-  using base_type =
-      KernelsAccessor<specfem::dimension::type::dim2, MediumTag,
-                      specfem::element::property_tag::isotropic_cosserat,
-                      UseSIMD>;
-
-  using value_type = typename base_type::value_type;
-  using simd = typename base_type::simd;
-
-  data_container() {
-    Kokkos::abort(
-        "Kernels container for elastic isotropic cosserat media is not "
-        "implemented for this dimension");
-  }
-};
-
-/** @} */ // end of group
-
-/**
- * @defgroup specfem_point_kernels_dim2_acoustic_isotropic 2D Acoustic Kernels
- * @{
- */
-
-/**
- * @ingroup specfem_point_kernels_dim2_acoustic_isotropic
- * @brief Data container to hold misfit kernels of 2D acoustic media at a
- * quadrature point
- *
- * @tparam UseSIMD Boolean indicating whether to use SIMD intrinsics
- *
- * <b> Accessor functions generated by the </b> \c POINT_CONTAINER \c MACRO
- *
- * - Get @f$ K_{\rho} @f$
- *   @code KOKKOS_INLINE_FUNCTION const value_type rho() const @endcode
- * - Get @f$ K_{\kappa} @f$
- *   @code KOKKOS_INLINE_FUNCTION const value_type kappa() const @endcode
- * - Get @f$ K_{\rho_p} @f$
- *   @code KOKKOS_INLINE_FUNCTION const value_type rhop() const @endcode
- * - Get @f$ K_{\alpha} @f$
- *   @code KOKKOS_INLINE_FUNCTION const value_type alpha() const @endcode
- */
-template <bool UseSIMD>
-struct data_container<specfem::dimension::type::dim2,
-                      specfem::element::medium_tag::acoustic,
-                      specfem::element::property_tag::isotropic, UseSIMD>
-    : public KernelsAccessor<specfem::dimension::type::dim2,
-                             specfem::element::medium_tag::acoustic,
-                             specfem::element::property_tag::isotropic,
-                             UseSIMD> {
-
-  using base_type =
-      KernelsAccessor<specfem::dimension::type::dim2,
-                      specfem::element::medium_tag::acoustic,
-                      specfem::element::property_tag::isotropic, UseSIMD>;
-
-  using value_type = typename base_type::value_type;
-  using simd = typename base_type::simd;
-
-  POINT_CONTAINER(rho, kappa, rhop, alpha)
-
-  KOKKOS_FUNCTION
-  data_container(const value_type rho, const value_type kappa)
-      : data_container(rho, kappa, rho * kappa,
-                       static_cast<type_real>(2.0) * kappa) {}
-};
-
-/** @} */ // end of group
-
-/**
- * @defgroup specfem_point_kernels_dim2_electromagnetic 2D Electromagnetic
- * Kernels
- * @{
- */
-
-/**
- * @ingroup specfem_point_kernels_dim2_electromagnetic
- * @brief Data container to hold misfit kernels of 2D electromagnetic media at a
- * quadrature point
- *
- * @tparam MediumTag The medium tag
- * @tparam UseSIMD Boolean indicating whether to use SIMD intrinsics
- */
-template <specfem::element::medium_tag MediumTag, bool UseSIMD>
-struct data_container<
-    specfem::dimension::type::dim2, MediumTag,
-    specfem::element::property_tag::isotropic, UseSIMD,
-    std::enable_if_t<specfem::element::is_electromagnetic<MediumTag>::value> >
-    : public KernelsAccessor<specfem::dimension::type::dim2, MediumTag,
-                             specfem::element::property_tag::isotropic,
-                             UseSIMD> {
-  using base_type =
-      KernelsAccessor<specfem::dimension::type::dim2, MediumTag,
-                      specfem::element::property_tag::isotropic, UseSIMD>;
-
-  using value_type = typename base_type::value_type;
-  using simd = typename base_type::simd;
-
-  KOKKOS_FUNCTION
-  data_container() {
-    Kokkos::abort("Kernels container for electromagnetic media is not "
-                  "implemented for this dimension");
-  }
-};
-
-/** @} */ // end of group
-
-/**
- * @defgroup specfem_point_kernels_dim2_poroelastic_isotropic 2D Poroelastic
- * Kernels
- * @{
- */
-
-/**
- * @ingroup specfem_point_kernels_dim2_poroelastic_isotropic
- * @brief Data container to hold misfit kernels of 2D poroelastic media at a
- * quadrature point
- *
- * @tparam UseSIMD Boolean indicating whether to use SIMD intrinsics
- *
- * <b> Accessor functions generated by the </b> \c POINT_CONTAINER \c MACRO
- *
- * - Get @f$ K_{\rho_t} @f$
- *   @code KOKKOS_INLINE_FUNCTION const value_type rhot() const @endcode
- * - Get @f$ K_{\rho_f} @f$
- *   @code KOKKOS_INLINE_FUNCTION const value_type rhof() const @endcode
- * - Get @f$ K_{\eta} @f$
- *   @code KOKKOS_INLINE_FUNCTION const value_type eta() const @endcode
- * - Get @f$ K_{s_m} @f$
- *   @code KOKKOS_INLINE_FUNCTION const value_type sm() const @endcode
- * - Get @f$ K_{\mu_{fr}} @f$
- *   @code KOKKOS_INLINE_FUNCTION const value_type mu_fr() const @endcode
- * - Get @f$ K_{B} @f$
- *   @code KOKKOS_INLINE_FUNCTION const value_type B() const @endcode
- * - Get @f$ K_{C} @f$
- *   @code KOKKOS_INLINE_FUNCTION const value_type C() const @endcode
- * - Get @f$ K_{M} @f$
- *   @code KOKKOS_INLINE_FUNCTION const value_type M() const @endcode
- * - Get @f$ K_{\mu_{frb}} @f$
- *   @code KOKKOS_INLINE_FUNCTION const value_type mu_frb() const @endcode
- * - Get @f$ K_{\rho_b} @f$
- *   @code KOKKOS_INLINE_FUNCTION const value_type rhob() const @endcode
- * - Get @f$ K_{\rho_{fb}} @f$
- *   @code KOKKOS_INLINE_FUNCTION const value_type rhofb() const @endcode
- * - Get @f$ K_{\phi} @f$
- *   @code KOKKOS_INLINE_FUNCTION const value_type phi() const @endcode
- * - Get @f$ K_{cpI} @f$
- *   @code KOKKOS_INLINE_FUNCTION const value_type cpI() const @endcode
- * - Get @f$ K_{cpII} @f$
- *   @code KOKKOS_INLINE_FUNCTION const value_type cpII() const @endcode
- * - Get @f$ K_{cs} @f$
- *   @code KOKKOS_INLINE_FUNCTION const value_type cs() const @endcode
- * - Get @f$ K_{\rho_{bb}} @f$
- *   @code KOKKOS_INLINE_FUNCTION const value_type rhobb() const @endcode
- * - Get @f$ K_{\rho_{fbb}} @f$
- *   @code KOKKOS_INLINE_FUNCTION const value_type rhofbb() const @endcode
- * - Get @f$ K_{ratio} @f$
- *   @code KOKKOS_INLINE_FUNCTION const value_type ratio() const @endcode
- * - Get @f$ K_{\phi_b} @f$
- *   @code KOKKOS_INLINE_FUNCTION const value_type phib() const @endcode
- */
-template <bool UseSIMD>
-struct data_container<specfem::dimension::type::dim2,
-                      specfem::element::medium_tag::poroelastic,
-                      specfem::element::property_tag::isotropic, UseSIMD>
-    : public KernelsAccessor<specfem::dimension::type::dim2,
-                             specfem::element::medium_tag::poroelastic,
-                             specfem::element::property_tag::isotropic,
-                             UseSIMD> {
-
-  using base_type =
-      KernelsAccessor<specfem::dimension::type::dim2,
-                      specfem::element::medium_tag::poroelastic,
-                      specfem::element::property_tag::isotropic, UseSIMD>;
-  using value_type = typename base_type::value_type;
-  using simd = typename base_type::simd;
-
-  POINT_CONTAINER(rhot, rhof, eta, sm, mu_fr, B, C, M, mu_frb, rhob, rhofb, phi,
-                  cpI, cpII, cs, rhobb, rhofbb, ratio, phib)
-
-  KOKKOS_FUNCTION
-  data_container(const value_type rhot, const value_type rhof,
-                 const value_type eta, const value_type sm,
-                 const value_type mu_fr, const value_type B, const value_type C,
-                 const value_type M, const value_type cpI,
-                 const value_type cpII, const value_type cs,
-                 const value_type rhobb, const value_type rhofbb,
-                 const value_type ratio, const value_type phib)
-      : data_container(rhot, rhof, eta, sm, mu_fr, B, C, M, mu_fr,
-                       (rhot + B + mu_fr), (rhof + C + M + sm),
-                       (static_cast<value_type>(-1.0) * (sm + M)), cpI, cpII,
-                       cs, rhobb, rhofbb, ratio, phib) {}
-};
-
-/** @} */ // end of group
-
-} // namespace kernels
-
-} // namespace impl
+namespace specfem::point {
 
 template <specfem::dimension::type DimensionTag,
           specfem::element::medium_tag MediumTag,
@@ -344,5 +28,4 @@ struct kernels : public impl::kernels::data_container<DimensionTag, MediumTag,
   constexpr static auto property_tag = PropertyTag; ///< type of the properties
 };
 
-} // namespace point
-} // namespace specfem
+} // namespace specfem::point
