+++ include/medium/dim2/acoustic/isotropic/kernels.hpp
@@ -1,0 +1,58 @@
+#pragma once
+
+#include "medium/impl/point_container.hpp"
+#include <Kokkos_Core.hpp>
+
+namespace specfem::point::impl::kernels {
+
+/**
+ * @defgroup specfem_point_kernels_dim2_acoustic_isotropic 2D Acoustic Kernels
+ * @{
+ */
+
+/**
+ * @ingroup specfem_point_kernels_dim2_acoustic_isotropic
+ * @brief Data container to hold misfit kernels of 2D acoustic media at a
+ * quadrature point
+ *
+ * @tparam UseSIMD Boolean indicating whether to use SIMD intrinsics
+ *
+ * <b> Accessor functions generated by the </b> \c POINT_CONTAINER \c MACRO
+ *
+ * - Get @f$ K_{\rho} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type rho() const @endcode
+ * - Get @f$ K_{\kappa} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type kappa() const @endcode
+ * - Get @f$ K_{\rho_p} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type rhop() const @endcode
+ * - Get @f$ K_{\alpha} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type alpha() const @endcode
+ */
+template <bool UseSIMD>
+struct data_container<specfem::dimension::type::dim2,
+                      specfem::element::medium_tag::acoustic,
+                      specfem::element::property_tag::isotropic, UseSIMD>
+    : public KernelsAccessor<specfem::dimension::type::dim2,
+                             specfem::element::medium_tag::acoustic,
+                             specfem::element::property_tag::isotropic,
+                             UseSIMD> {
+
+  using base_type =
+      KernelsAccessor<specfem::dimension::type::dim2,
+                      specfem::element::medium_tag::acoustic,
+                      specfem::element::property_tag::isotropic, UseSIMD>;
+
+  using value_type = typename base_type::value_type;
+  using simd = typename base_type::simd;
+
+  POINT_CONTAINER(rho, kappa, rhop, alpha)
+
+  KOKKOS_FUNCTION
+  data_container(const value_type rho, const value_type kappa)
+      : data_container(rho, kappa, rho * kappa,
+                       static_cast<type_real>(2.0) * kappa) {}
+};
+
+/** @} */ // end of group
+
+} // namespace specfem::point::impl::kernels
