+++ include/medium/dim2/elastic/anisotropic/kernels.hpp
@@ -1,0 +1,58 @@
+#pragma once
+
+#include "medium/impl/point_container.hpp"
+#include <Kokkos_Core.hpp>
+
+namespace specfem::point::impl::kernels {
+
+/**
+ * @defgroup specfem_point_kernels_dim2_elastic_anisotropic 2D Elastic
+ * Anisotropic Kernels
+ * @{
+ */
+
+/**
+ * @ingroup specfem_point_kernels_dim2_elastic_anisotropic
+ * @brief Data container to hold misfit kernels of 2D elastic anisotropic media
+ * at a quadrature point
+ *
+ * @tparam UseSIMD Boolean indicating whether to use SIMD intrinsics
+ *
+ * <b> Accessor functions generated by the </b> \c POINT_CONTAINER \c MACRO
+ *
+ * - Get @f$ K_{\rho} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type rho() const @endcode
+ * - Get @f$ K_{c_{11}} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type c11() const @endcode
+ * - Get @f$ K_{c_{13}} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type c13() const @endcode
+ * - Get @f$ K_{c_{15}} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type c15() const @endcode
+ * - Get @f$ K_{c_{33}} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type c33() const @endcode
+ * - Get @f$ K_{c_{35}} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type c35() const @endcode
+ * - Get @f$ K_{c_{55}} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type c55() const @endcode
+ */
+template <specfem::element::medium_tag MediumTag, bool UseSIMD>
+struct data_container<
+    specfem::dimension::type::dim2, MediumTag,
+    specfem::element::property_tag::anisotropic, UseSIMD,
+    std::enable_if_t<specfem::element::is_elastic<MediumTag>::value> >
+    : public KernelsAccessor<specfem::dimension::type::dim2, MediumTag,
+                             specfem::element::property_tag::anisotropic,
+                             UseSIMD> {
+  using base_type =
+      KernelsAccessor<specfem::dimension::type::dim2, MediumTag,
+                      specfem::element::property_tag::anisotropic, UseSIMD>;
+
+  using value_type = typename base_type::value_type;
+  using simd = typename base_type::simd;
+
+  POINT_CONTAINER(rho, c11, c13, c15, c33, c35, c55)
+};
+
+/** @} */ // end of group
+
+} // namespace specfem::point::impl::kernels
