+++ include/medium/dim2/elastic/isotropic/kernels.hpp
@@ -1,0 +1,56 @@
+#pragma once
+
+#include "medium/impl/point_container.hpp"
+#include <Kokkos_Core.hpp>
+
+namespace specfem::point::impl::kernels {
+
+/**
+ * @defgroup specfem_point_kernels_elastic_isotropic 2D Elastic Isotropic
+ * Kernels
+ * @{
+ */
+/**
+ * @ingroup specfem_point_kernels_elastic_isotropic
+ * @brief Data container to hold misfit kernels of 2D elastic isotropic media at
+ * a quadrature point
+ *
+ * @tparam UseSIMD Boolean indicating whether to use SIMD intrinsics
+ *
+ * <b> Accessor functions generated by the </b> \c POINT_CONTAINER \c MACRO
+ *
+ * - Get @f$ K_{\rho} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type rho() const @endcode
+ * - Get @f$ K_{\mu} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type mu() const @endcode
+ * - Get @f$ K_{\kappa} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type kappa() const @endcode
+ * - Get @f$ K_{\rho^\prime} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type rhop() const @endcode
+ * - Get @f$ K_{\alpha} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type alpha() const @endcode
+ * - Get @f$ K_{\beta} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type beta() const @endcode
+ *
+ */
+template <specfem::dimension::type DimensionTag,
+          specfem::element::medium_tag MediumTag, bool UseSIMD>
+struct data_container<
+    DimensionTag, MediumTag, specfem::element::property_tag::isotropic, UseSIMD,
+    std::enable_if_t<specfem::element::is_elastic<MediumTag>::value> >
+    : public KernelsAccessor<DimensionTag, MediumTag,
+                             specfem::element::property_tag::isotropic,
+                             UseSIMD> {
+
+  using base_type =
+      KernelsAccessor<DimensionTag, MediumTag,
+                      specfem::element::property_tag::isotropic, UseSIMD>;
+
+  using value_type = typename base_type::value_type;
+  using simd = typename base_type::simd;
+
+  POINT_CONTAINER(rho, mu, kappa, rhop, alpha, beta)
+};
+/** @} */ // end of group
+
+} // namespace specfem::point::impl::kernels
