+++ include/medium/dim2/elastic/isotropic_cosserat/kernels.hpp
@@ -1,0 +1,46 @@
+#pragma once
+
+#include "medium/impl/point_container.hpp"
+#include <Kokkos_Core.hpp>
+
+namespace specfem::point::impl::kernels {
+
+/**
+ * @defgroup specfem_point_kernels_dim2_elastic_isotropic_cosserat 2D Cosserat
+ * Kernels
+ * @{
+ */
+
+/**
+ * @brief Placeholder container for 2D elastic isotropic cosserat media
+ *
+ * @tparam MediumTag
+ * @tparam UseSIMD
+ */
+template <specfem::element::medium_tag MediumTag, bool UseSIMD>
+struct data_container<
+    specfem::dimension::type::dim2, MediumTag,
+    specfem::element::property_tag::isotropic_cosserat, UseSIMD,
+    std::enable_if_t<specfem::element::is_elastic<MediumTag>::value> >
+    : public KernelsAccessor<specfem::dimension::type::dim2, MediumTag,
+                             specfem::element::property_tag::isotropic_cosserat,
+                             UseSIMD> {
+
+  using base_type =
+      KernelsAccessor<specfem::dimension::type::dim2, MediumTag,
+                      specfem::element::property_tag::isotropic_cosserat,
+                      UseSIMD>;
+
+  using value_type = typename base_type::value_type;
+  using simd = typename base_type::simd;
+
+  data_container() {
+    Kokkos::abort(
+        "Kernels container for elastic isotropic cosserat media is not "
+        "implemented for this dimension");
+  }
+};
+
+/** @} */ // end of group
+
+} // namespace specfem::point::impl::kernels
