+++ include/medium/dim2/poroelastic/isotropic/kernels.hpp
@@ -1,0 +1,97 @@
+#pragma once
+
+#include "medium/impl/point_container.hpp"
+#include <Kokkos_Core.hpp>
+
+namespace specfem::point::impl::kernels {
+
+/**
+ * @defgroup specfem_point_kernels_dim2_poroelastic_isotropic 2D Poroelastic
+ * Kernels
+ * @{
+ */
+
+/**
+ * @ingroup specfem_point_kernels_dim2_poroelastic_isotropic
+ * @brief Data container to hold misfit kernels of 2D poroelastic media at a
+ * quadrature point
+ *
+ * @tparam UseSIMD Boolean indicating whether to use SIMD intrinsics
+ *
+ * <b> Accessor functions generated by the </b> \c POINT_CONTAINER \c MACRO
+ *
+ * - Get @f$ K_{\rho_t} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type rhot() const @endcode
+ * - Get @f$ K_{\rho_f} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type rhof() const @endcode
+ * - Get @f$ K_{\eta} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type eta() const @endcode
+ * - Get @f$ K_{s_m} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type sm() const @endcode
+ * - Get @f$ K_{\mu_{fr}} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type mu_fr() const @endcode
+ * - Get @f$ K_{B} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type B() const @endcode
+ * - Get @f$ K_{C} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type C() const @endcode
+ * - Get @f$ K_{M} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type M() const @endcode
+ * - Get @f$ K_{\mu_{frb}} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type mu_frb() const @endcode
+ * - Get @f$ K_{\rho_b} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type rhob() const @endcode
+ * - Get @f$ K_{\rho_{fb}} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type rhofb() const @endcode
+ * - Get @f$ K_{\phi} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type phi() const @endcode
+ * - Get @f$ K_{cpI} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type cpI() const @endcode
+ * - Get @f$ K_{cpII} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type cpII() const @endcode
+ * - Get @f$ K_{cs} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type cs() const @endcode
+ * - Get @f$ K_{\rho_{bb}} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type rhobb() const @endcode
+ * - Get @f$ K_{\rho_{fbb}} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type rhofbb() const @endcode
+ * - Get @f$ K_{ratio} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type ratio() const @endcode
+ * - Get @f$ K_{\phi_b} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type phib() const @endcode
+ */
+template <bool UseSIMD>
+struct data_container<specfem::dimension::type::dim2,
+                      specfem::element::medium_tag::poroelastic,
+                      specfem::element::property_tag::isotropic, UseSIMD>
+    : public KernelsAccessor<specfem::dimension::type::dim2,
+                             specfem::element::medium_tag::poroelastic,
+                             specfem::element::property_tag::isotropic,
+                             UseSIMD> {
+
+  using base_type =
+      KernelsAccessor<specfem::dimension::type::dim2,
+                      specfem::element::medium_tag::poroelastic,
+                      specfem::element::property_tag::isotropic, UseSIMD>;
+  using value_type = typename base_type::value_type;
+  using simd = typename base_type::simd;
+
+  POINT_CONTAINER(rhot, rhof, eta, sm, mu_fr, B, C, M, mu_frb, rhob, rhofb, phi,
+                  cpI, cpII, cs, rhobb, rhofbb, ratio, phib)
+
+  KOKKOS_FUNCTION
+  data_container(const value_type rhot, const value_type rhof,
+                 const value_type eta, const value_type sm,
+                 const value_type mu_fr, const value_type B, const value_type C,
+                 const value_type M, const value_type cpI,
+                 const value_type cpII, const value_type cs,
+                 const value_type rhobb, const value_type rhofbb,
+                 const value_type ratio, const value_type phib)
+      : data_container(rhot, rhof, eta, sm, mu_fr, B, C, M, mu_fr,
+                       (rhot + B + mu_fr), (rhof + C + M + sm),
+                       (static_cast<value_type>(-1.0) * (sm + M)), cpI, cpII,
+                       cs, rhobb, rhofbb, ratio, phib) {}
+};
+
+/** @} */ // end of group
+
+} // namespace specfem::point::impl::kernels
